Carry a cast's target as a field, and resolve relation qualifiers and NULL struct members
Cast and TryCast now hold the field they yield rather than a bare type. A
type-only target (Cast::new) keeps the source's metadata, less any Arrow
extension type; a target that carries metadata states the output's identity
exactly, so a cast to a type whose identity lives in metadata yields that
identity whatever the source carried. The physical CastExpr follows the same
rule through cast_output_field, and an unclaimed SQL cast yields the written
target field.

A wildcard qualifier resolves the way a column reference does: `t.*` names
`s.t` when no other relation of the schema answers to `t`.

Union type resolution gives a NULL struct or list member the type it meets,
as it already did for a top-level NULL, so rows that differ only in where
their NULLs sit share one type.

Expression planners can claim a call written with named arguments
(plan_named_call, given the written argument order) and an IS [NOT] DISTINCT
FROM whose operands the null-safe comparison does not compare, such as rows
(plan_distinct_from).

+++ datafusion/sql/src/expr/substring.rs
@@ -140,7 +140,9 @@ impl SqlToRel<'_> {
                     schema,
                     planner_context,
                 )?;
-                let from_logic = Expr::Literal(ScalarValue::Int64(Some(1)), None);
+                // `SUBSTRING(text FOR n)` starts at the integer constant 1,
+                // the start PostgreSQL's grammar supplies for the form.
+                let from_logic = Expr::Literal(ScalarValue::Int32(Some(1)), None);
                 let for_logic = self.sql_expr_to_logical_expr(
                     for_expr.as_ref(),
                     schema,
